requirements.txt
Package                  Version
------------------------ -----------
certifi                  2024.8.30
chardet                  3.0.4
charset-normalizer       3.3.2
click                    8.1.7
easyocr                  1.7.1
filelock                 3.16.0
fsspec                   2024.9.0
googletrans              4.0.0rc1
h11                      0.9.0
h2                       3.2.0
hpack                    3.0.0
hstspreload              2024.9.1
httpcore                 0.9.1
httpx                    0.13.3
hyperframe               5.2.0
idna                     2.10
imageio                  2.35.1
Jinja2                   3.1.4
lazy_loader              0.4
libretranslatepy         2.1.1
lxml                     5.3.0
MarkupSafe               2.1.5
mpmath                   1.3.0
networkx                 3.3
ninja                    1.11.1.1
numpy                    2.1.1
nvidia-cublas-cu12       12.1.3.1
nvidia-cuda-cupti-cu12   12.1.105
nvidia-cuda-nvrtc-cu12   12.1.105
nvidia-cuda-runtime-cu12 12.1.105
nvidia-cudnn-cu12        9.1.0.70
nvidia-cufft-cu12        11.0.2.54
nvidia-curand-cu12       10.3.2.106
nvidia-cusolver-cu12     11.4.5.107
nvidia-cusparse-cu12     12.1.0.106
nvidia-nccl-cu12         2.20.5
nvidia-nvjitlink-cu12    12.6.68
nvidia-nvtx-cu12         12.1.105
opencv-python            4.10.0.84
opencv-python-headless   4.10.0.84
packaging                24.1
pillow                   10.4.0
pip                      24.2
pyclipper                1.3.0.post5
python-bidi              0.6.0
PyYAML                   6.0.2
requests                 2.32.3
rfc3986                  1.5.0
scikit-image             0.24.0
scipy                    1.14.1
setuptools               74.1.2
shapely                  2.0.6
sniffio                  1.3.1
sympy                    1.13.2
tifffile                 2024.8.30
torch                    2.4.1
torchvision              0.19.1
translate                3.6.1
triton                   3.0.0
typing_extensions        4.12.2
urllib3                  2.2.2
